+++ packages/pangea-node-sdk/src/request.ts
@@ -181,7 +181,7 @@ class PangeaRequest {
     this.postPresignedURL(presigned_url, {
       file: fileData.file,
       file_details: file_details,
-      name: fileData.name,
+      name: "file",
     });
     return response.gotResponse;
   }