+++ packages/pangea-node-sdk/src/services/index.ts
@@ -12,6 +12,7 @@ import {
 } from "./intel.js";
 import VaultService from "./vault.js";
 import { FileScanService } from "./file_scan.js";
+import SanitizeService from "./sanitize.js";
 
 export default {
   AuditService,
@@ -26,4 +27,5 @@ export default {
   UserIntelService,
   VaultService,
   FileScanService,
+  SanitizeService,
 };